% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_soi.R
\name{clean_soi}
\alias{clean_soi}
\title{helper function to clean IRS SOI data}
\usage{
clean_soi(irs_raw)
}
\arguments{
\item{irs_raw}{Uncleaned data.table of annual tax data}
}
\description{
Cleans raw IRS annual data from \code{\link[=load_soi]{load_soi()}} by removing summary rows,
aggregating income levels and coalescing number of return columns,
which vary from early to later years and number of return columns
}
