% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_agi.R
\name{convert_agi}
\alias{convert_agi}
\title{Helper function used by load_soi() to normalize pre 2008 levels}
\usage{
convert_agi(agi_level)
}
\arguments{
\item{agi_level}{A State ID Code}
}
\description{
Helper to aggregate income levels from multiple year's of data
}
