% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_agi_2.R
\name{convert_agi_2}
\alias{convert_agi_2}
\title{Helper function used by load_soi() to normalize post 2008 levels}
\usage{
convert_agi_2(agi_level)
}
\arguments{
\item{agi_level}{A State ID Code}
}
\description{
Helper to aggregate income levels from multiple year's of data
}
