% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_agi_3.R
\name{convert_agi_3}
\alias{convert_agi_3}
\title{Helper function used by load_soi() to name agi_stub}
\usage{
convert_agi_3(agi_stub)
}
\arguments{
\item{agi_stub}{A State ID Code}
}
\description{
Helper to aggregate income levels from multiple year's of data
}
