% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dict}
\alias{data_dict}
\title{Data dictionary}
\format{
A data frame with 968 rows and 3 variables:
\describe{
\item{names}{IRS SOI Field, chr}
\item{row_lists}{Field Description, chr}
\item{year}{year, chr}
}
}
\source{
\url{https://www.nber.org/research/data/individual-income-tax-statistics-zip-code-data-soi/}
}
\usage{
data_dict
}
\description{
Data dictionary
}
\keyword{datasets}
