% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nber_data.R
\name{download_nber_data}
\alias{download_nber_data}
\title{Download data from NBER and store in designated folder}
\usage{
download_nber_data(path, start_year, finish_year)
}
\arguments{
\item{path}{"/home/irs_data/"}

\item{start_year}{2005}

\item{finish_year}{2016}
}
\description{
Given years to download and folder path, downloads raw .csv data
(https://www.nber.org/research/data/individual-income-tax-statistics-zip-code-data-soi)
Download time is approximately 1 minute per year
}
\examples{
\dontrun{library(data.table)
 download_nber_data(start = 2016)}

}
