% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_load_prepare_soi.R
\name{duckdb_load_prepare_soi}
\alias{duckdb_load_prepare_soi}
\title{Load and Prepare NBER SOI Data}
\usage{
duckdb_load_prepare_soi(states = "", path = "")
}
\arguments{
\item{states}{A State ID Code}

\item{path}{The location of the .csv folder output of download_nber_data()}
}
\description{
Given folder of NBER Downloaded annual .csv's from \code{\link[=download_nber_data]{download_nber_data()}}, return a data.table
(https://www.nber.org/research/data/individual-income-tax-statistics-zip-code-data-soi)
}
\details{
\itemize{
\item Example of state would be "CT" or "FL"
\item Directory path used for download_nber_data() (ie: "/home/..")
}
}
\examples{
\dontrun{
library(duckplyr)
irs <- duckdb_load_prepare_soi("~/Documents/Data/irs_zip_soi/")}

}
