% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_lookup.R
\name{field_lookup}
\alias{field_lookup}
\title{Lookup Field Description from IRS SOI Data}
\usage{
field_lookup(select_year = "2016", select_name = "a00100")
}
\arguments{
\item{select_year}{Int year}

\item{select_name}{Field identifier}
}
\description{
Takes year and NBER Field identifier and returns Field Description
(https://www.nber.org/research/data/individual-income-tax-statistics-zip-code-data-soi)
}
\details{
\itemize{
\item Example is year = "2016" and name = "a00100"
}
}
\examples{
\dontrun{library(data.table)
field_lookup("2016", "a00100")}

}
