% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_classes.R
\name{get_col_classes}
\alias{get_col_classes}
\title{Get column class types for use in load_soi()}
\usage{
get_col_classes(file)
}
\arguments{
\item{file}{A .csv file loaded from the IRS SOI folder}
}
\description{
Given a .csv, determine the column types based on regex matches
}
