% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_row_dt.R
\name{get_summary_row_dt}
\alias{get_summary_row_dt}
\title{Helper function to receive irs data.table and return annual summary rows for that zipcode}
\usage{
get_summary_row_dt(irs)
}
\arguments{
\item{irs}{IRS data.table}
}
\description{
Receive full dataset and return only summary rows usually denoted "00000" or "99999"
}
