% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_agg_calc.R
\name{group_agg_calc}
\alias{group_agg_calc}
\title{Aggregates Group Amount}
\usage{
group_agg_calc(data, soi_number = "a00100", soi_count = "n1")
}
\arguments{
\item{data}{IRS data.table}

\item{soi_number}{Income or deduction chr value}

\item{soi_count}{Returns reporting income or deduction chr value}
}
\description{
Takes full data and aggregates by requested group variable with inputs
}
