% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irs_app_data}
\alias{irs_app_data}
\title{Clean IRS Data for Shiny App}
\format{
An object of class \code{data.frame} with 2033904 rows and 9 columns.
}
\usage{
irs_app_data
}
\description{
Cleaned up IRS Data
}
\examples{
irs_app_data

}
\keyword{datasets}
