% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irsApp.R
\name{irsApp}
\alias{irsApp}
\title{Run Shiny App with available data}
\usage{
irsApp(data = irs_app_data)
}
\arguments{
\item{data}{A data.table of IRS SOI data}
}
\description{
Takes available IRS data and runs Shiny App for exploration
(https://www.nber.org/research/data/individual-income-tax-statistics-zip-code-data-soi)
}
\examples{
\dontrun{
irsApp()}

}
