% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_soi.R
\name{load_soi}
\alias{load_soi}
\title{Import IRS SOI NBER Data}
\usage{
load_soi(state = "", path = path)
}
\arguments{
\item{state}{A State ID Code}

\item{path}{The location of the .csv folder}
}
\description{
Given folder of NBER Downloaded annual .csv's from \code{\link[=download_nber_data]{download_nber_data()}}, return a data.table
(https://www.nber.org/research/data/individual-income-tax-statistics-zip-code-data-soi)
}
\details{
\itemize{
\item Example of state would be "CT" or "FL"
\item Path should be "/home/.."
}
}
\examples{
\dontrun{library(data.table)
irs <- load_soi("/home/irs_data/")}

}
