% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_agi_graph.R
\name{make_agi_graph}
\alias{make_agi_graph}
\title{Plotly graph of per capita items}
\usage{
make_agi_graph(data)
}
\arguments{
\item{data}{IRS data.table}
}
\description{
Takes cleaned IRS data.table and makes Plotly graph of aggregated income divided by returns by selection
}
\examples{
\dontrun{
make_agi_graph(irs)}

}
