% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_deductions_DT.R
\name{make_deductions_DT}
\alias{make_deductions_DT}
\title{Function to take incoming reactive data.frame, convert to data.table and build Summary DT}
\usage{
make_deductions_DT(data)
}
\arguments{
\item{data}{IRS data.table}
}
\description{
Makes a Summary datatable of AGI, Federal Tax, Total returns and Unique Zips by year
using output of \code{\link[=clean_soi]{clean_soi()}}
}
