% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spaghetti_plot.R
\name{make_spaghetti_plot}
\alias{make_spaghetti_plot}
\title{Plotly graph of comparing selected components}
\usage{
make_spaghetti_plot(data)
}
\arguments{
\item{data}{IRS data.table}
}
\description{
Takes cleaned IRS data.table and makes Plotly graph of aggregated income divided by returns by selection
using output of \code{\link[=clean_soi]{clean_soi()}}
}
\examples{
\dontrun{
make_spaghetti_plot(irs_app_data, "zipcode")}

}
