% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tax_graph.R
\name{make_tax_graph}
\alias{make_tax_graph}
\title{Plotly graph of per tax rate}
\usage{
make_tax_graph(data)
}
\arguments{
\item{data}{IRS data.table}
}
\description{
Takes cleaned IRS data.table and makes Plotly graph of tax rate by selection
}
\examples{
\dontrun{
make_tax_graph(irs)}

}
