% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_app_data.R
\name{prepare_app_data}
\alias{prepare_app_data}
\title{Function to select cols and add city/county to irs for use in \code{\link[=irsApp]{irsApp()}}}
\usage{
prepare_app_data(irs)
}
\arguments{
\item{irs}{Cleaned IRS data.table from \code{\link[=clean_soi]{clean_soi()}}}
}
\description{
Helper to
}
\examples{
\dontrun{library(data.table)
irs_app_data <- prepare_app_data(irs)}

}
