% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{Zipcodes}
\format{
A data frame with 33144 rows and 3 variables:
\describe{
\item{zipcode}{Zipcode, chr}
\item{county}{County, chr}
\item{major_city}{Major City, chr}
\item{land_area_in_sqmi}{Land Area, num}
\item{lat}{Latitude, num}
\item{lng}{Longitude, num}
}
@source \url{https://www.unitedstateszipcodes.org/zip-code-database/}
}
\usage{
zipcodes
}
\description{
Zipcodes
}
\keyword{datasets}
