% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irs_app_data}
\alias{irs_app_data}
\title{Clean IRS Data for Shiny App}
\format{
A data frame with 2033904 rows and 9 variables:
\describe{
\item{year}{year, int}
\item{zipcode}{zipcode chr}
\item{agi_level}{five income levels, chr}
\item{n1}{number of taxpayers, int}
\item{a00100}{Adjusted Gross Income, num}
\item{total_tax}{Total tax paid, num}
}
}
\source{
\url{https://www.nber.org/research/data/individual-income-tax-statistics-zip-code-data-soi/}
}
\usage{
irs_app_data
}
\description{
Sample of cleaned up IRS Data product after running load_soi() and clean_soi()
Only includes 2018 for CT
}
\keyword{datasets}
