% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_agi_graph.R
\name{make_agi_graph}
\alias{make_agi_graph}
\title{Plotly graph of per capita items}
\usage{
make_agi_graph(data, type = "agi")
}
\arguments{
\item{data}{IRS data.table}

\item{type}{chr vector to specify if agi or per_cap}
}
\description{
Takes cleaned IRS data.table and makes Plotly graph of aggregated income divided by returns by selection
using output of \code{\link[=prepare_app_data]{prepare_app_data()}}
}
\examples{
\dontrun{
make_agi_graph(irs)}

}
